pyproject.toml
[project]
name = "agent-tools"
version = "0.1.0"
description = "AI Agent's tool for Python package method analysis and API discovery"
authors = [{ name = "Graham Anderson", email = "graham@grahama.co" }]
requires-python = ">=3.8"
license = { text = "MIT" }
readme = "README.md"
dependencies = [
    # Core dependencies
    "loguru>=0.7.0",
    "tqdm>=4.66.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    # Cursor Rules dependencies
    "python-arango>=7.5.0",
    "transformers>=4.38.2",
    "sentence-transformers>=2.5.1",
    "torch>=2.0.0",
    "numpy>=1.21.0",
    "tabulate>=0.9.0",
    "click>=8.0.0",
    "spacy>=3.5.0",
    "nltk>=3.8.0",
    "litellm>=1.61.20",
    "pytest>=8.3.5",
    "pylint>=3.2.7",
    "einops>=0.8.1",
    "json-repair>=0.30.0",
    "psutil>=7.0.0",
    "pytest-asyncio>=0.24.0",
    "rapidfuzz>=3.9.7",
    "deepmerge>=2.0",
    "tenacity>=9.0.0",
    "pillow>=10.4.0",
    "redis>=5.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-asyncio>=0.24.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    # "cursor-rules @ git+https://github.com/grahama1970/snippets.git#egg=cursor-rules&subdirectory=.cursor/rules"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
#packages = ["find:src/agent_tools"]
packages = ["find:src"]


[tool.hatch.build.targets.wheel]
#packages = ["find:src/agent_tools"]
packages = ["find:src"]


[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 100
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
src_paths = ["src"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
mypy_path = "src"

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
src = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-n auto --dist loadfile"
testpaths = ["tests", "src/agent_tools"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[project.scripts]
method-validator = "agent_tools.method_validator.cli:main"
cursor-rules = "agent_tools.cursor_rules.cli:cli"

[tool.cursor.patterns]
pattern_directories = [".cursor/rules"]
install_location = "project_root"

# Cursor Rules configuration
[tool.cursor-rules]
auto_setup = true
db_url = "http://localhost:8529"
db_name = "cursor_rules"
rules_dir = ".cursor/rules"
enable_embeddings = true
enable_views = true
default_priority = 3
critical_categories = [
    "database",
    "async",
    "testing"
]
